+++ packages/device_info_plus/CHANGELOG.md
@@ -1,5 +1,6 @@
 ## NEXT
 
+* Update analysis_options.yaml for Flutter 3.47.0.
 * Add 2 integration test cases.
 * Update integration tests.