+++ packages/device_info_plus/CHANGELOG.md
@@ -1,6 +1,7 @@
-## NEXT
+## 1.4.1
 
 * Update analysis_options.yaml for Flutter 3.47.0.
+* Update the repository URL to use the `main` branch.
 * Add 2 integration test cases.
 * Update integration tests.